Show indeterminate progress while loading
Loading the HTML takes roughly half the time. If the output is
over 1000 pages, this is enough time for Windows to mark the program
as "not responding". Setting the progress bar style to marquee
lets the OS and the user know that something is happening.

Also, add Geckofx to the spelling dictionary for developers who care